compiles but very broken

+++ Crowny/Source/Crowny/Common/StringUtils.h
@@ -10,7 +10,7 @@ namespace Crowny
 		static int32_t ParseInt(const std::string& value);
 		static float ParseFloat(const std::string& value);
 		static uint64_t ParseLong(const std::string& value);
-		static uint64_t ParseDouble(const std::string& value);
+		static double ParseDouble(const std::string& value);
 		
 		static const char* FindToken(const char* string, const std::string& token);
